+++ packages/cli/src/helpers.ts
@@ -4,6 +4,7 @@ import {
   PsyoptionsEuropeanInstrument,
   LegInstrument,
   FixedSize,
+  TransactionPriority,
 } from '@convergence-rfq/sdk';
 import { Command } from 'commander';
 
@@ -52,6 +53,11 @@ export const formatInstrument = (instrument: Instrument): string => {
 export const addDefaultArgs = (cmd: any) => {
   cmd.option('--rpc-endpoint <string>', 'RPC endpoint', DEFAULT_RPC_ENDPOINT);
   cmd.option('--skip-preflight', 'skip preflight', false);
+  cmd.option(
+    '--tx-priority-fee <string>',
+    'transaction priority fee can be [none : 0 mcLamports , normal : 1 mcLamports, high : 10 mcLamports, turbo : 100 mcLamports, custom : <number> mcLamports]',
+    'none'
+  );
   cmd.option('--keypair-file <string>', 'keypair file', DEFAULT_KEYPAIR_FILE);
   cmd.option('--verbose <boolean>', 'verbose', false);
   return cmd;
@@ -143,4 +149,29 @@ export const fetchBirdeyeTokenPrice = async (
     return Number(tokenPriceJson?.data?.value);
   }
   return undefined;
+};
+
+export const resolveTxPriorityArg = (
+  txPriority: string
+): TransactionPriority => {
+  switch (txPriority) {
+    case 'none':
+      return 'none';
+    case 'normal':
+      return 'normal';
+    case 'high':
+      return 'high';
+    case 'turbo':
+      return 'turbo';
+    default:
+      try {
+        const txPriorityInNumber = Number(txPriority);
+        if (isNaN(txPriorityInNumber) || txPriorityInNumber < 0) {
+          return 'none';
+        }
+        return txPriorityInNumber;
+      } catch (e) {
+        return 'none';
+      }
+  }
 };