+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @convergence-rfq/cli
 
+## 6.3.1
+
+### Patch Changes
+
+- Add tx-priority-fee argument to cli and modify sdk txPriority to take custom values
+- Updated dependencies
+  - @convergence-rfq/sdk@6.3.1
+
 ## 6.3.0
 
 ### Minor Changes